requirements.txt
pytest
pytest-cov
pandas
python-dotenv